use shared loader for case export
